turn off spammer emails

+++ server/controller/forum/NodeCtrler.php
@@ -83,7 +83,7 @@ class NodeCtrler extends Forum
                }
             }
 
-            if ( $this->config->webmaster )
+            if ( FALSE && $this->config->webmaster ) // turn off spammer emails
             {
                $mailer = new Mailer();
                $mailer->subject = 'SPAMMER detected and deleted (' . \sizeof( $users ) . ($deleteAll ? ' deleted)' : ' not deleted)');
